# Conflicts:
#	server/main.go
#	server/sequences/sequences.go

+++ test/sequences/sequences.go
@@ -9,7 +9,7 @@ import (
 
 // on start waits for time to be hour o'clock
 // then starts chron routine that is timed on every 4 hours
-func SaveOnFourHoursPeriod(temp chan float32) {
+func SaveOnFourHoursPeriod(temp chan float64) {
 	//fmt.Println(temp)
 	// for time.Now().Format("04") != "00" {
 	// 	// TEST
@@ -26,9 +26,9 @@ func SaveOnFourHoursPeriod(temp chan float32) {
 	<-gocron.Start()
 }
 
-func MeasurementSequence(cTemp chan float32) {
+func MeasurementSequence(cTemp chan float64) {
 	gocron.Every(1).Seconds().Do(func() {
-		temperature := float32(rand.Float64() * 5)
+		temperature := float64(rand.Float64() * 5)
 
 		fmt.Println("Measure\nTemperature: %v˚C", temperature, cTemp)
